+++ stack.h
@@ -7,7 +7,7 @@
 struct stack {
     int top;
     int length;
-    int *data;
+    double *data;
 };
 
 struct stack * create_stack(int length);
@@ -16,9 +16,9 @@ void free_stack(struct stack *s);
 
 void push(struct stack *s, double value);
 
-int peek(struct stack *s);
+double peek(struct stack *s);
 
-int pop(struct stack *s);
+double pop(struct stack *s);
 
 void resize(struct stack *s, int new_length);
